Split simplified_circuit out of gate_decomposition Second draft of qec_inverse
