+++ file_lists/source_files_no_main
@@ -92,3 +92,4 @@ src/stim/util_bot/error_decomp.cc
 src/stim/util_top/circuit_inverse_unitary.cc
 src/stim/util_top/circuit_to_detecting_regions.cc
 src/stim/util_top/circuit_vs_amplitudes.cc
+src/stim/util_top/simplified_circuit.cc
