+++ src/stim/util_top/simplified_circuit.h
@@ -1,0 +1,12 @@
+#ifndef _STIM_UTIL_TOP_SIMPLIFIED_CIRCUIT_H
+#define _STIM_UTIL_TOP_SIMPLIFIED_CIRCUIT_H
+
+#include "stim/circuit/circuit.h"
+
+namespace stim {
+
+Circuit simplified_circuit(const Circuit &circuit);
+
+}  // namespace stim
+
+#endif
